ddtrace/tracer/span: rename method LocalRootSpan into Root
