+++ ddtrace/tracer/span.go
@@ -170,15 +170,18 @@ func (s *span) setSamplingPriority(priority int, sampler samplernames.SamplerNam
 	s.setSamplingPriorityLocked(priority, sampler)
 }
 
-// LocalRootSpan returns the root span of the span's trace.
-func (s *span) LocalRootSpan() Span {
-	return s.localRootSpan()
+// Root returns the root span of the span's trace. The return value shouldn't be
+// nil as long as the root span is valid and not finished.
+func (s *span) Root() Span {
+	return s.root()
 }
 
-// localRootSpan returns the root span of the span's trace. As opposed to the public LocalRootSpan method, this
-// returns the actual span type when internal usage requires it (to avoid type assertions from LocalRootSpan's return
+// root returns the root span of the span's trace. The return value shouldn't be
+// nil as long as the root span is valid and not finished.
+// As opposed to the public Root method, this one returns the actual span type
+// when internal usage requires it (to avoid type assertions from Root's return
 // value).
-func (s *span) localRootSpan() *span {
+func (s *span) root() *span {
 	if s == nil || s.context == nil {
 		return nil
 	}
@@ -198,7 +201,7 @@ func (s *span) SetUser(id string, opts ...UserMonitoringOption) {
 	for _, fn := range opts {
 		fn(&cfg)
 	}
-	root := s.localRootSpan()
+	root := s.root()
 	trace := root.context.trace
 	root.Lock()
 	defer root.Unlock()
